created a new hook, used useUnstakeReportingStake hook in details report
